backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.2.0"
description = "backend system for a specialized CRM for real estate agencies"
authors = ["Alireza Amini <alireza.4mini@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "5"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
celery = {extras = ["redis"], version = "^5.4.0"}
django-taggit = "^6.1.0"
django-extensions = "^3.2.3"
django-compressor = "^4.5.1"
django-cors-headers = "^4.6.0"
requests = "^2.32.3"
djangorestframework-simplejwt = "^5.3.1"
jdatetime = "^5.0.0"
pytz = "^2024.2"
pillow = "^11.0.0"
whitenoise = "^6.8.2"
gunicorn = "^23.0.0"
pytest = "^8.3.3"
pytest-django = "^4.9.0"
redis = "^5.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
